OPSK-317 - auth lookup table for new user
fix to correctly update the lookup table after a new user has been associated with an existing profile. Previously the job was triggered when the user model was created, but this is too soon and before being linked to the project. Now fixed to update after the link to the profile.